Main method problem and Packaging
