+++ src/main/java/com/mpather47/git/HospitalManagementSystem.java
@@ -3,11 +3,10 @@ package com.mpather47.git;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+
+@SpringBootApplication
 public class HospitalManagementSystem {
-    @SpringBootApplication
-    public static class Example {
         public static void main(String[] args) {
-            SpringApplication.run(Example.class, args);
+            SpringApplication.run(HospitalManagementSystem.class,args);
         }
-    }
 }
