+++ src/main/java/com/mpather47/git/services/appointment/AppointmentService.java
@@ -3,7 +3,6 @@ package com.mpather47.git.services.appointment;
 import com.mpather47.git.entity.appointment.Appointment;
 import com.mpather47.git.services.IService;
 
-import javax.xml.ws.Service;
 import java.util.Set;
 
 public interface AppointmentService extends IService<Appointment, String> {
