+++ src/test/java/com/mpather47/git/controller/hospital/HospitalControllerTest.java
@@ -1,9 +1,8 @@
-package com.mpather47.git.controller;
+package com.mpather47.git.controller.hospital;
 
-import com.mpather47.git.entity.person.Address;
-import com.mpather47.git.entity.person.Person;
-import com.mpather47.git.factory.person.AddressFactory;
-import com.mpather47.git.factory.person.PersonFactory;
+
+import com.mpather47.git.entity.hospital.Hospital;
+import com.mpather47.git.factory.hospital.HospitalFactory;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -23,46 +22,46 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @RunWith(SpringRunner.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class AddressControllerTest {
-    private static Address address = AddressFactory.addAddress("5454",234234,PersonFactory.createPerson("424234","324324"));
+public class HospitalControllerTest {
+    private static Hospital hospital = HospitalFactory.createHospital("23453", "St Mary", 1254);
     @Autowired
     private TestRestTemplate restTemplate;
-    private String baseURL = "http://localhost:8080/address/";
+    private String baseURL = "http://localhost:8080/hospital/";
 
     @Test
     public void a_create(){
         String url = baseURL + "create/";
         System.out.println("URL: " + url);
-        System.out.println("Post data: " + address);
-        ResponseEntity<Address> postResponse = restTemplate.postForEntity(url,address,Address.class);
+        System.out.println("Post data: " + hospital);
+        ResponseEntity<Hospital> postResponse = restTemplate.postForEntity(url,hospital,Hospital.class);
         assertNotNull(postResponse);
         assertNotNull(postResponse.getBody());
-        address = postResponse.getBody();
-        System.out.println("Saved data: " + address);
-        assertEquals(address.getAddressId(),postResponse.getBody().getAddressId());
+        hospital = postResponse.getBody();
+        System.out.println("Saved data: " + hospital);
+        assertEquals(hospital.getHospitalId(),postResponse.getBody().getHospitalId());
     }
 
     @Test
     public void b_read(){
-        String url = baseURL + "read/" +address.getAddressId();
+        String url = baseURL + "read/" + hospital.getHospitalId();
         System.out.println("URL: " + url);
-        ResponseEntity<Address> response = restTemplate.getForEntity(url,Address.class);
-        assertEquals(address.getAddressId(),response.getBody().getAddressId());
+        ResponseEntity<Hospital> response = restTemplate.getForEntity(url,Hospital.class);
+        assertEquals(hospital.getHospitalId(),response.getBody().getHospitalId());
     }
 
     @Test
     public void c_update(){
-        Address updated = new Address.Builder().copy(address).setPostcode(34234).setAddress("545345").setPostcode(25435543).build();
+        Hospital updated = new Hospital.Builder().copy(hospital).setHospitalName("st mary").setHospitalNumber(323).build();
         String url = baseURL + "update/";
         System.out.println("Post data:" + updated);
-        ResponseEntity<Address> response = restTemplate.postForEntity(url,updated, Address.class);
+        ResponseEntity<Hospital> response = restTemplate.postForEntity(url, updated, Hospital.class);
         System.out.println(response);
-        assertEquals(address.getAddressId(),response.getBody().getAddressId());
+        assertEquals(updated.getHospitalId(),response.getBody().getHospitalId());
     }
 
     @Test
     public void e_delete(){
-        String url = baseURL + "delete/" + address.getAddressId();
+        String url = baseURL + "delete/" + hospital.getHospitalId();
         System.out.println("URL: " + url);
         restTemplate.delete(url);
     }
@@ -79,3 +78,4 @@ public class AddressControllerTest {
     }
 
 }
+
