+++ src/test/java/com/mpather47/git/controller/person/AddressControllerTest.java
@@ -1,7 +1,8 @@
-package com.mpather47.git.controller;
+package com.mpather47.git.controller.person;
 
 import com.mpather47.git.entity.person.Address;
 import com.mpather47.git.entity.person.Person;
+import com.mpather47.git.factory.person.AddressFactory;
 import com.mpather47.git.factory.person.PersonFactory;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
@@ -22,48 +23,46 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @RunWith(SpringRunner.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class PersonControllerTest {
-
-    private static Person person = PersonFactory.createPerson("Marcell","1998/08/21");
-
+public class AddressControllerTest {
+    private static Address address = AddressFactory.addAddress("5454",234234,PersonFactory.createPerson("424234","324324"));
     @Autowired
     private TestRestTemplate restTemplate;
-    private String baseURL = "http://localhost:8080/person/";
+    private String baseURL = "http://localhost:8080/address/";
 
     @Test
     public void a_create(){
-        String url = baseURL + "create";
-        System.out.println("URL" + url);
-        System.out.printf("Post data: " + person);
-        ResponseEntity<Person> posResponse = restTemplate.postForEntity(url, person, Person.class);
-        assertNotNull(posResponse);
-        assertNotNull(posResponse.getBody());
-        person = posResponse.getBody();
-        System.out.println("Saved data: " + person);
-        assertEquals(person.getPersonId(),posResponse.getBody().getPersonId());
+        String url = baseURL + "create/";
+        System.out.println("URL: " + url);
+        System.out.println("Post data: " + address);
+        ResponseEntity<Address> postResponse = restTemplate.postForEntity(url,address,Address.class);
+        assertNotNull(postResponse);
+        assertNotNull(postResponse.getBody());
+        address = postResponse.getBody();
+        System.out.println("Saved data: " + address);
+        assertEquals(address.getAddressId(),postResponse.getBody().getAddressId());
     }
 
     @Test
     public void b_read(){
-        String url = baseURL +  "read/" + person.getPersonId();
+        String url = baseURL + "read/" +address.getAddressId();
         System.out.println("URL: " + url);
-        ResponseEntity<Person> response = restTemplate.getForEntity(url,Person.class);
-        assertEquals(person.getPersonId(),response.getBody().getPersonId());
+        ResponseEntity<Address> response = restTemplate.getForEntity(url,Address.class);
+        assertEquals(address.getAddressId(),response.getBody().getAddressId());
     }
 
     @Test
     public void c_update(){
-        Person updated = new Person.Builder().copy(person).setName("gsgfgd").setDateOfBirth("fgsdgsd").build();
+        Address updated = new Address.Builder().copy(address).setPostcode(34234).setAddress("545345").setPostcode(25435543).build();
         String url = baseURL + "update/";
         System.out.println("Post data:" + updated);
-        ResponseEntity<Person> response = restTemplate.postForEntity(url,updated, Person.class);
+        ResponseEntity<Address> response = restTemplate.postForEntity(url,updated, Address.class);
         System.out.println(response);
-        assertEquals(person.getPersonId(),response.getBody().getPersonId());
+        assertEquals(address.getAddressId(),response.getBody().getAddressId());
     }
 
     @Test
     public void e_delete(){
-        String url = baseURL + "delete/" + person.getPersonId();
+        String url = baseURL + "delete/" + address.getAddressId();
         System.out.println("URL: " + url);
         restTemplate.delete(url);
     }
@@ -78,6 +77,5 @@ public class PersonControllerTest {
         System.out.println(response.getBody());
 
     }
-
 
 }
