+++ src/test/java/com/mpather47/git/controller/person/ContactControllerTest.java
@@ -1,9 +1,10 @@
-package com.mpather47.git.controller;
+package com.mpather47.git.controller.person;
 
-
-import com.mpather47.git.entity.hospital.Equipment;
-import com.mpather47.git.entity.hospital.Room;
-import com.mpather47.git.factory.hospital.EquipmentFactory;
+import com.mpather47.git.entity.person.Address;
+import com.mpather47.git.entity.person.Contact;
+import com.mpather47.git.entity.person.Person;
+import com.mpather47.git.factory.person.ContactFactory;
+import com.mpather47.git.factory.person.PersonFactory;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -16,6 +17,10 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -23,47 +28,48 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @RunWith(SpringRunner.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class EquipmentControllerTest {
-    private static Room a;
-    private static Equipment equipment = EquipmentFactory.createEquipment("232", "Scalpel", "Cutting tool", 5,  a);
+public class ContactControllerTest {
+    private static Contact contact = ContactFactory.addContact(234432,2342434,"4234",PersonFactory.createPerson("Marcel Pather","1241344"));
+
     @Autowired
     private TestRestTemplate restTemplate;
-    private String baseURL = "http://localhost:8080/equipment/";
+    private String baseURL = "http://localhost:8080/contact/";
 
     @Test
     public void a_create(){
         String url = baseURL + "create/";
         System.out.println("URL: " + url);
-        System.out.println("Post data: " + equipment);
-        ResponseEntity<Equipment> postResponse = restTemplate.postForEntity(url, equipment,Equipment.class);
+        System.out.println("Post data: " + contact);
+        ResponseEntity<Contact> postResponse = restTemplate.postForEntity(url,contact,Contact.class);
         assertNotNull(postResponse);
         assertNotNull(postResponse.getBody());
-        equipment = postResponse.getBody();
-        System.out.println("Saved data: " + equipment);
-        assertEquals(equipment.getEquipmentId(),postResponse.getBody().getEquipmentId());
+        contact = postResponse.getBody();
+        System.out.println("Saved data: " + contact);
+        assertEquals(contact.getContactId(),postResponse.getBody().getContactId());
     }
 
     @Test
     public void b_read(){
-        String url = baseURL + "read/" + equipment.getEquipmentId();
+        String url = baseURL + "read/" + contact.getContactId();
         System.out.println("URL: " + url);
-        ResponseEntity<Equipment> response = restTemplate.getForEntity(url,Equipment.class);
-        assertEquals(equipment.getEquipmentId(),response.getBody().getEquipmentId());
+        ResponseEntity<Contact> response = restTemplate.getForEntity(url,Contact.class);
+        assertEquals(contact.getContactId(),response.getBody().getContactId());
+
     }
 
     @Test
     public void c_update(){
-        Equipment updated = new Equipment.Builder().copy(equipment).setName("knife").setDesc("another cutting tool").setQuantity(3).setDetails(a).build();
+        Contact updated = new Contact.Builder().copy(contact).setEmail("ggdf").setCellphoneNo(2434324).setHomePhone(23434342).build();
         String url = baseURL + "update/";
         System.out.println("Post data:" + updated);
-        ResponseEntity<Equipment> response = restTemplate.postForEntity(url, updated, Equipment.class);
+        ResponseEntity<Contact> response = restTemplate.postForEntity(url,updated, Contact.class);
         System.out.println(response);
-        assertEquals(equipment.getEquipmentId(),response.getBody().getEquipmentId());
+        assertEquals(contact.getContactId(),response.getBody().getContactId());
     }
 
     @Test
     public void e_delete(){
-        String url = baseURL + "delete/" + equipment.getEquipmentId();
+        String url = baseURL + "delete/" + contact.getContactId();
         System.out.println("URL: " + url);
         restTemplate.delete(url);
     }
@@ -80,5 +86,3 @@ public class EquipmentControllerTest {
     }
 
 }
-
-
