+++ src/test/java/com/mpather47/git/controller/person/PersonControllerTest.java
@@ -1,9 +1,7 @@
-package com.mpather47.git.controller;
+package com.mpather47.git.controller.person;
 
 import com.mpather47.git.entity.person.Address;
-import com.mpather47.git.entity.person.Contact;
 import com.mpather47.git.entity.person.Person;
-import com.mpather47.git.factory.person.ContactFactory;
 import com.mpather47.git.factory.person.PersonFactory;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
@@ -17,10 +15,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -28,48 +22,48 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @RunWith(SpringRunner.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class ContactControllerTest {
-    private static Contact contact = ContactFactory.addContact(234432,2342434,"4234",PersonFactory.createPerson("Marcel Pather","1241344"));
+public class PersonControllerTest {
+
+    private static Person person = PersonFactory.createPerson("Marcell","1998/08/21");
 
     @Autowired
     private TestRestTemplate restTemplate;
-    private String baseURL = "http://localhost:8080/contact/";
+    private String baseURL = "http://localhost:8080/person/";
 
     @Test
     public void a_create(){
-        String url = baseURL + "create/";
-        System.out.println("URL: " + url);
-        System.out.println("Post data: " + contact);
-        ResponseEntity<Contact> postResponse = restTemplate.postForEntity(url,contact,Contact.class);
-        assertNotNull(postResponse);
-        assertNotNull(postResponse.getBody());
-        contact = postResponse.getBody();
-        System.out.println("Saved data: " + contact);
-        assertEquals(contact.getContactId(),postResponse.getBody().getContactId());
+        String url = baseURL + "create";
+        System.out.println("URL" + url);
+        System.out.printf("Post data: " + person);
+        ResponseEntity<Person> posResponse = restTemplate.postForEntity(url, person, Person.class);
+        assertNotNull(posResponse);
+        assertNotNull(posResponse.getBody());
+        person = posResponse.getBody();
+        System.out.println("Saved data: " + person);
+        assertEquals(person.getPersonId(),posResponse.getBody().getPersonId());
     }
 
     @Test
     public void b_read(){
-        String url = baseURL + "read/" + contact.getContactId();
+        String url = baseURL +  "read/" + person.getPersonId();
         System.out.println("URL: " + url);
-        ResponseEntity<Contact> response = restTemplate.getForEntity(url,Contact.class);
-        assertEquals(contact.getContactId(),response.getBody().getContactId());
-
+        ResponseEntity<Person> response = restTemplate.getForEntity(url,Person.class);
+        assertEquals(person.getPersonId(),response.getBody().getPersonId());
     }
 
     @Test
     public void c_update(){
-        Contact updated = new Contact.Builder().copy(contact).setEmail("ggdf").setCellphoneNo(2434324).setHomePhone(23434342).build();
+        Person updated = new Person.Builder().copy(person).setName("gsgfgd").setDateOfBirth("fgsdgsd").build();
         String url = baseURL + "update/";
         System.out.println("Post data:" + updated);
-        ResponseEntity<Contact> response = restTemplate.postForEntity(url,updated, Contact.class);
+        ResponseEntity<Person> response = restTemplate.postForEntity(url,updated, Person.class);
         System.out.println(response);
-        assertEquals(contact.getContactId(),response.getBody().getContactId());
+        assertEquals(person.getPersonId(),response.getBody().getPersonId());
     }
 
     @Test
     public void e_delete(){
-        String url = baseURL + "delete/" + contact.getContactId();
+        String url = baseURL + "delete/" + person.getPersonId();
         System.out.println("URL: " + url);
         restTemplate.delete(url);
     }
@@ -84,5 +78,6 @@ public class ContactControllerTest {
         System.out.println(response.getBody());
 
     }
+
 
 }
